requirements.txt
Click==7.0
Flask==1.1.1
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
python-dotenv==0.10.5
Werkzeug==0.15.6
yapf==0.29.0
pyrebase==3.0.27
flask-recaptcha==0.4.2
flake8==3.7.9
pytest==5.4.1
gcloud==0.17.0
google-api-core==1.22.2
google-api-python-client==1.12.1
google-auth==1.21.2
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.51.0